Split Case Study editing into multiple edit pages with subnavigation
