+++ app/Providers/Filament/AppPanelProvider.php
@@ -110,6 +110,6 @@ class AppPanelProvider extends PanelProvider
                             ]),
                     ]);
             })
-        ;
+            ->unsavedChangesAlerts();
     }
 }
